+++ cookbooks/arcgis-enterprise/resources/patches.rb
@@ -2,7 +2,7 @@
 # Cookbook Name:: arcgis-enterprise
 # Resource:: patches
 #
-# Copyright 2018 Esri
+# Copyright 2022 Esri
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,6 +18,13 @@
 #
 
 actions :install
+
+attribute :patch_folder, :kind_of => String
+attribute :patch, :kind_of => String
+attribute :product, :kind_of => String
+attribute :run_as_user, :kind_of => String
+attribute :patch_log, :kind_of => String
+attribute :patch_registry, :kind_of => String
 
 def initialize(*args)
   super