+++ cookbooks/arcgis-geoevent/resources/geoevent.rb
@@ -2,7 +2,7 @@
 # Cookbook Name:: arcgis-geoevent
 # Resource:: geoevent
 #
-# Copyright 2015 Esri
+# Copyright 2022 Esri
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -23,6 +23,7 @@ actions :system, :unpack, :install, :uninstall, :update_account,
 attribute :setup_archive, :kind_of => String
 attribute :setups_repo, :kind_of => String
 attribute :setup, :kind_of => String
+attribute :setup_options, :kind_of => String, :default => ''
 attribute :install_dir, :kind_of => String
 attribute :run_as_user, :kind_of => String
 attribute :run_as_password, :kind_of => String