+++ cookbooks/esri-iis/metadata.rb
@@ -4,7 +4,8 @@ maintainer_email 'contracts@esri.com'
 license          'Apache 2.0'
 description      'Enables IIS features required by ArcGIS Web Adaptor (IIS) and configures HTTPS binding.'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          '0.1.3'
+version          '0.2.0'
+chef_version     '>= 14.0' if defined? chef_version
 
 depends          'windows', '~> 5.3'
 depends          'openssl', '~> 8.5'