+++ cookbooks/arcgis-pro/CHANGELOG.md
@@ -3,6 +3,11 @@ arcgis-pro cookbook CHANGELOG
 
 This file is used to list changes made in each version of the arcgis-pro cookbook.
 
+4.0.0
+-----
+- Added support for ArcGIS Pro 3.0.
+- Support for installing .NET 6 Desktop Runtime x64
+
 3.8.0
 -----
 - Added support for ArcGIS Pro 2.8/2.9.